ConceptDoesNotExistException moved
